gdemon-xen: release checkpoint
In this patch, we present gDemon, which is a concrete implementation of
Demon. gDemon reuses the guest GTT as the 1st-dimensional translation
and utilizes the system IOMMU to construct a 2nd-dimensional address
translation. gDemon dynamically switches the 2nd-dimensional page table
when the device owner switches. In order to enable guest VM to access
its own global graphics memory, such as AGP Aperture, even when the
corresponding vGPU is not current device owner, GGTT remains
partitioned, shadowed, and shifted to a dedicated address range that
PPGTTs won't map.
The evaluations show that gDemon provides up to 19.73x better
performance in the media transcoding workloads and achieves the
performance improvement of up to 17.09% and 13.73% in 2D benchmarks and
3D benchmarks, respectively, compared to gVirt. Moreover, gDemon
simplifies the implementation of GPU MMU virtualization by 37% code
reduction.

In detail, we have made the following changes in xen.
* set up a separated I/O page table for GPU device
    - remap GGTT entry from HPA + 128G to HPA
* implement switch_gpu_iopt hypercall
    - switch I/O page table among all VM candidates
* implement lookup_gpu_addr hypercall
    - get vtd mapping from gfn to mfn
* partital flush iotlb of GPU device
    - set the did field in gpu context entry to nr_dom - 1
    - only flush iotlb in the range of [0, MAX_GUEST_MEMORY] using psi
* others
    - support multi-domain bootup and shutdown
    - remove a lot of help functions used in tricky shortpath.